Add additional null check for getHeading() within FilamentShield->getLocalizedWidgetLabel()
This check is necessary because Filament\Widgets\StatsOverviewWidget.php always has a base getHeading() method.

This causes the "default" branch within the match statement to not execute.

This currently works as-is for the table widgets because it uses a different method name --> getTableHeading() and does not have a getHeading() method.

Adding this null check should give us the expected behavior.